# openapi5.yaml
...
openapi: 3.0.3
info:
  title: TaskMind API
  description: REST API менеджера задач и базы знаний ANotes/TaskMind
  version: 5.0.0
paths:
...
  /api/project-statuses/project/{projectId}:
    get:
      tags:
      - Project Status Resource
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectStatusResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - bearerAuth: []
  /api/projects:
    get:
      tags:
      - Project Resource
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - bearerAuth: []
    post:
      tags:
      - Project Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequest"
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - bearerAuth: []
...
  /api/projects/{id}:
    get:
      tags:
      - Project Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - bearerAuth: []
    delete:
      tags:
      - Project Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - bearerAuth: []
    patch:
      tags:
      - Project Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - bearerAuth: []
  /api/projects/{id}/tags:
    get:
      tags:
      - Project Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
...
    ProjectRequest:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 50
          minLength: 3
          pattern: \S
          type: string
        description:
          type: string
        color:
          pattern: "#[0-9A-Fa-f]{6}"
          type: string
        icon:
          maxLength: 50
          type: string
        tags:
          maxItems: 30
          type: array
          items:
            type: string
    ProjectResponse:
      type: object
      properties:
        id:
          format: int32
          type: integer
        name:
          type: string
        description:
          type: string
        ownerUserId:
          format: int32
          type: integer
        color:
          type: string
        icon:
          type: string
        tags:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
...
    ProjectStatusResponse:
      type: object
      properties:
        id:
          format: int32
          type: integer
        projectId:
          format: int32
          type: integer
        statusName:
          type: string
        statusColor:
          type: string
        statusOrder:
          format: int32
          type: integer
        isDefault:
          type: boolean
        isClosed:
          type: boolean
...
    ProjectUpdateRequest:
      type: object
      properties:
        name:
          maxLength: 50
          minLength: 3
          type: string
        description:
          type: string
        color:
          pattern: "#[0-9A-Fa-f]{6}"
          type: string
        icon:
          maxLength: 50
          type: string
        tags:
          maxItems: 30
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT из /api/auth/login или /api/auth/register
      scheme: bearer
      bearerFormat: JWT